+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 # Change Log
 
 ## 0.9.19
-
+- Added configuration option to control the max cache size while resolving references in semantic analysis, defaulting to 500MB
 
 ## 0.9.18
 - Fixed a rare case where the DLS would crash when reporting device contexts